requirements.txt
Django
djangorestframework
python-environ
psycopg2
django-filter